+++ apps/utils/image_io.cpp
@@ -49,8 +49,7 @@ OIDN_NAMESPACE_BEGIN
 
     std::shared_ptr<ImageBuffer> loadImagePFM(const DeviceRef& device,
                                               const std::string& filename,
-                                              int numChannels,
-                                              Format dataType)
+                                              DataType dataType)
     {
       // Open the file
       std::ifstream file(filename, std::ios::binary);
@@ -68,13 +67,8 @@ OIDN_NAMESPACE_BEGIN
       else
         throw std::runtime_error("invalid PFM image");
 
-      if (numChannels == 0)
-        numChannels = C;
-      else if (C < numChannels)
-        throw std::runtime_error("not enough image channels");
-
-      if (dataType == Format::Undefined)
-        dataType = Format::Float;
+      if (dataType == DataType::Void)
+        dataType = DataType::Float32;
 
       int H, W;
       file >> W >> H;
@@ -92,7 +86,7 @@ OIDN_NAMESPACE_BEGIN
       scale = fabs(scale);
 
       // Read the pixels
-      auto image = std::make_shared<ImageBuffer>(device, W, H, numChannels, dataType);
+      auto image = std::make_shared<ImageBuffer>(device, W, H, C, dataType);
 
       for (int h = 0; h < H; ++h)
       {
@@ -102,8 +96,8 @@ OIDN_NAMESPACE_BEGIN
           {
             float x;
             file.read((char*)&x, sizeof(float));
-            if (c < numChannels)
-              image->set((size_t(H-1-h)*W + w) * numChannels + c, x * scale);
+            if (c < C)
+              image->set((size_t(H-1-h)*W + w) * C + c, x * scale);
           }
         }
       }
@@ -154,8 +148,7 @@ OIDN_NAMESPACE_BEGIN
 
     std::shared_ptr<ImageBuffer> loadImagePHM(const DeviceRef& device,
                                               const std::string& filename,
-                                              int numChannels,
-                                              Format dataType)
+                                              DataType dataType)
     {
       // Open the file
       std::ifstream file(filename, std::ios::binary);
@@ -173,13 +166,8 @@ OIDN_NAMESPACE_BEGIN
       else
         throw std::runtime_error("invalid PHM image");
 
-      if (numChannels == 0)
-        numChannels = C;
-      else if (C < numChannels)
-        throw std::runtime_error("not enough image channels");
-
-      if (dataType == Format::Undefined)
-        dataType = Format::Half;
+      if (dataType == DataType::Void)
+        dataType = DataType::Float16;
 
       int H, W;
       file >> W >> H;
@@ -197,7 +185,7 @@ OIDN_NAMESPACE_BEGIN
       scale = fabs(scale);
 
       // Read the pixels
-      auto image = std::make_shared<ImageBuffer>(device, W, H, numChannels, dataType);
+      auto image = std::make_shared<ImageBuffer>(device, W, H, C, dataType);
 
       for (int h = 0; h < H; ++h)
       {
@@ -207,17 +195,14 @@ OIDN_NAMESPACE_BEGIN
           {
             half x;
             file.read((char*)&x, sizeof(x));
-            if (c < numChannels)
+            if (scale == 1.f)
             {
-              if (scale == 1.f)
-              {
-                image->set((size_t(H-1-h)*W + w) * numChannels + c, x);
-              }
-              else
-              {
-                const float xs = float(x) * scale;
-                image->set((size_t(H-1-h)*W + w) * numChannels + c, xs);
-              }
+              image->set((size_t(H-1-h)*W + w) * C + c, x);
+            }
+            else
+            {
+              const float xs = float(x) * scale;
+              image->set((size_t(H-1-h)*W + w) * C + c, xs);
             }
           }
         }
@@ -300,24 +285,21 @@ OIDN_NAMESPACE_BEGIN
   #ifdef OIDN_USE_OPENIMAGEIO
     std::shared_ptr<ImageBuffer> loadImageOIIO(const DeviceRef& device,
                                                const std::string& filename,
-                                               int numChannels,
-                                               Format dataType)
+                                               DataType dataType)
     {
       auto in = OIIO::ImageInput::open(filename);
       if (!in)
         throw std::runtime_error("cannot open image file: '" + filename + "'");
 
       const OIIO::ImageSpec& spec = in->spec();
-      if (numChannels == 0)
-        numChannels = spec.nchannels;
-      else if (spec.nchannels < numChannels)
-        throw std::runtime_error("not enough image channels");
+      const int numChannels = std::min(spec.nchannels, 3);
 
-      if (dataType == Format::Undefined)
-        dataType = (spec.channelformat(0) == OIIO::TypeDesc::HALF) ? Format::Half : Format::Float;
+      if (dataType == DataType::Void)
+        dataType = (spec.channelformat(0) == OIIO::TypeDesc::HALF) ? DataType::Float16 : DataType::Float32;
 
       auto image = std::make_shared<ImageBuffer>(device, spec.width, spec.height, numChannels, dataType);
-      bool success = in->read_image(0, 0, 0, numChannels, dataType == Format::Half ? OIIO::TypeDesc::HALF : OIIO::TypeDesc::FLOAT, image->getData());
+      bool success = in->read_image(0, 0, 0, numChannels,
+        dataType == DataType::Float16 ? OIIO::TypeDesc::HALF : OIIO::TypeDesc::FLOAT, image->getData());
       in->close();
 
   #if OIIO_VERSION < 10903
@@ -339,10 +321,10 @@ OIDN_NAMESPACE_BEGIN
       OIIO::TypeDesc format;
       switch (image.getDataType())
       {
-      case Format::Float:
+      case DataType::Float32:
         format = OIIO::TypeDesc::FLOAT;
         break;
-      case Format::Half:
+      case DataType::Float16:
         format = OIIO::TypeDesc::HALF;
         break;
       default:
@@ -372,19 +354,18 @@ OIDN_NAMESPACE_BEGIN
 
   std::shared_ptr<ImageBuffer> loadImage(const DeviceRef& device,
                                          const std::string& filename,
-                                         int numChannels,
-                                         Format dataType)
+                                         DataType dataType)
   {
     const std::string ext = getExtension(filename);
     std::shared_ptr<ImageBuffer> image;
 
     if (ext == "pfm")
-      image = loadImagePFM(device, filename, numChannels, dataType);
+      image = loadImagePFM(device, filename, dataType);
     else if (ext == "phm")
-      image = loadImagePHM(device, filename, numChannels, dataType);
+      image = loadImagePHM(device, filename, dataType);
     else
 #if OIDN_USE_OPENIMAGEIO
-      image = loadImageOIIO(device, filename, numChannels, dataType);
+      image = loadImageOIIO(device, filename, dataType);
 #else
       throw std::runtime_error("cannot load unsupported image file format: '" + filename + "'");
 #endif
@@ -417,11 +398,10 @@ OIDN_NAMESPACE_BEGIN
 
   std::shared_ptr<ImageBuffer> loadImage(const DeviceRef& device,
                                          const std::string& filename,
-                                         int numChannels,
                                          bool srgb,
-                                         Format dataType)
+                                         DataType dataType)
   {
-    auto image = loadImage(device, filename, numChannels, dataType);
+    auto image = loadImage(device, filename, dataType);
     if (!srgb && isSrgbImage(filename))
       srgbInverse(*image);
     return image;