+++ common/common.cpp
@@ -9,9 +9,9 @@ OIDN_NAMESPACE_BEGIN
   {
     switch (dataType)
     {
-    case DataType::Float32: return sizeof(float);
-    case DataType::Float16: return sizeof(int16_t);
     case DataType::UInt8:   return 1;
+    case DataType::Float16: return sizeof(int16_t);
+    case DataType::Float32: return sizeof(float);
     default:
       throw std::invalid_argument("invalid data type");
     }
@@ -21,6 +21,8 @@ OIDN_NAMESPACE_BEGIN
   {
     switch (format)
     {
+    case Format::Undefined:
+      return DataType::Void;
     case Format::Float:
     case Format::Float2:
     case Format::Float3:
@@ -34,6 +36,30 @@ OIDN_NAMESPACE_BEGIN
     default:
       throw std::invalid_argument("invalid format");
     }
+  }
+
+  Format makeFormat(DataType dataType, int numChannels)
+  {
+    if (dataType == DataType::Void)
+      return Format::Undefined;
+
+    Format baseFormat;
+    switch (dataType)
+    {
+    case DataType::Float16:
+      baseFormat = Format::Half;
+      break;
+    case DataType::Float32:
+      baseFormat = Format::Float;
+      break;
+    default:
+      throw std::invalid_argument("unsupported format data type");
+    }
+
+    if (numChannels < 1 || numChannels > 4)
+      throw std::invalid_argument("invalid number of channels");
+
+    return Format(int(baseFormat) + numChannels - 1);
   }
 
 OIDN_NAMESPACE_END